Cleaning and reorganizing mostly Java side code

+++ languages/java/base/src/main/java/com/github/tengi/buffer/AbstractMemoryBuffer.java
@@ -52,9 +52,9 @@ public abstract class AbstractMemoryBuffer
 
     private volatile int lockCounter = 0;
 
-    protected long writerIndex = 0;
+    protected int writerIndex = 0;
 
-    protected long readerIndex = 0;
+    protected int readerIndex = 0;
 
     @Override
     public void lock()
@@ -123,7 +123,7 @@ public abstract class AbstractMemoryBuffer
     }
 
     @Override
-    public long readableBytes()
+    public int readableBytes()
     {
         return writerIndex() - readerIndex();
     }
@@ -159,7 +159,7 @@ public abstract class AbstractMemoryBuffer
     @Override
     public int readBuffer( ByteBuffer byteBuffer )
     {
-        int remaining = Math.min( byteBuffer.remaining(), (int) readableBytes() );
+        int remaining = Math.min( byteBuffer.remaining(), readableBytes() );
         return readBuffer( byteBuffer, byteBuffer.position(), remaining );
     }
 
@@ -181,14 +181,14 @@ public abstract class AbstractMemoryBuffer
     }
 
     @Override
-    public long readBuffer( WritableMemoryBuffer memoryBuffer )
+    public int readBuffer( WritableMemoryBuffer memoryBuffer )
     {
-        long remaining = Math.min( memoryBuffer.writableBytes(), readableBytes() );
+        int remaining = Math.min( memoryBuffer.writableBytes(), readableBytes() );
         return readBuffer( memoryBuffer, memoryBuffer.writerIndex(), remaining );
     }
 
     @Override
-    public long readBuffer( WritableMemoryBuffer memoryBuffer, long offset, long length )
+    public int readBuffer( WritableMemoryBuffer memoryBuffer, int offset, int length )
     {
         if ( memoryBuffer instanceof AbstractMemoryBuffer )
         {
@@ -200,9 +200,9 @@ public abstract class AbstractMemoryBuffer
         }
         else
         {
-            long mark = memoryBuffer.writerIndex();
+            int mark = memoryBuffer.writerIndex();
             memoryBuffer.writerIndex( offset );
-            for ( long pos = offset; pos < offset + length; pos++ )
+            for ( int pos = offset; pos < offset + length; pos++ )
             {
                 memoryBuffer.writeByte( readByte() );
             }
@@ -285,13 +285,13 @@ public abstract class AbstractMemoryBuffer
     }
 
     @Override
-    public long readerIndex()
+    public int readerIndex()
     {
         return readerIndex;
     }
 
     @Override
-    public void readerIndex( long readerIndex )
+    public void readerIndex( int readerIndex )
     {
         this.readerIndex = readerIndex;
     }
@@ -315,7 +315,7 @@ public abstract class AbstractMemoryBuffer
     }
 
     @Override
-    public long writableBytes()
+    public int writableBytes()
     {
         return maxCapacity() - writerIndex();
     }
@@ -338,7 +338,7 @@ public abstract class AbstractMemoryBuffer
     @Override
     public void writeBuffer( ByteBuffer byteBuffer )
     {
-        int remaining = Math.min( byteBuffer.remaining(), (int) writableBytes() );
+        int remaining = Math.min( byteBuffer.remaining(), writableBytes() );
         writeBuffer( byteBuffer, byteBuffer.position(), remaining );
     }
 
@@ -361,12 +361,12 @@ public abstract class AbstractMemoryBuffer
     @Override
     public void writeBuffer( ReadableMemoryBuffer memoryBuffer )
     {
-        long remaining = Math.min( memoryBuffer.readableBytes(), writableBytes() );
+        int remaining = Math.min( memoryBuffer.readableBytes(), writableBytes() );
         writeBuffer( memoryBuffer, memoryBuffer.readerIndex(), remaining );
     }
 
     @Override
-    public void writeBuffer( ReadableMemoryBuffer memoryBuffer, long offset, long length )
+    public void writeBuffer( ReadableMemoryBuffer memoryBuffer, int offset, int length )
     {
         if ( memoryBuffer instanceof AbstractMemoryBuffer )
         {
@@ -378,9 +378,9 @@ public abstract class AbstractMemoryBuffer
         }
         else
         {
-            long mark = memoryBuffer.readerIndex();
+            int mark = memoryBuffer.readerIndex();
             memoryBuffer.readerIndex( offset );
-            for ( long pos = offset; pos < offset + length; pos++ )
+            for ( int pos = offset; pos < offset + length; pos++ )
             {
                 writeByte( memoryBuffer.readByte() );
             }
@@ -462,13 +462,13 @@ public abstract class AbstractMemoryBuffer
     }
 
     @Override
-    public long writerIndex()
+    public int writerIndex()
     {
         return writerIndex;
     }
 
     @Override
-    public void writerIndex( long writerIndex )
+    public void writerIndex( int writerIndex )
     {
         this.writerIndex = writerIndex;
     }
